Group assessments after introduction
Proposing the first self assessment in the groups of 3 that are created at the end of the introduction. A relatively large number of small group assessment activities (~2 questions, 3-5 minutes) help participants identify points of confusion early and give them an opportunity to frequently access the information they're seeing and reinforce it.